+++ R/obscureCounts.R
@@ -47,8 +47,11 @@ obscureCounts <- function(table,
       checkColNames <- c("n_records", "n_person")
     } else if (grepl("missingValues", tableName)) {
       checkColNames <- c("n_records", "n_person", "n_records_not_missing_value", "n_records_missing_value")
-    } else if (grepl("diagnosticsSummary|conceptSummary", tableName)) {
+    } else if (grepl("conceptSummary", tableName)) {
       checkColNames <- colNames <- c("n_records", "n_patients")
+    } else if (grepl("diagnosticsSummary", tableName)) {
+      checkColNames <- c("n_records", "n_patients")
+      colNames <- c("n_records", "n_patients")
     }
 
     # if any count col is less than minCellCount and larger than zero, replace colNames with substitute