+++ NEWS.md
@@ -1,5 +1,9 @@
 # DrugExposureDiagnostics (development version)
 
+# DrugExposureDiagnostics 1.1.7
+* Suppress results diagnostics summary (#332)
+* Update ingredient concept plot section (#338)
+
 # DrugExposureDiagnostics 1.1.6
 * Make compatible with newest dplyr
 * Remove pryr dependency